'优化代码'

+++ src/main/java/com/gznx/service/UploadService.java
@@ -1,5 +1,6 @@
 package com.gznx.service;
 
+import com.gznx.domain.UploadFileInfo;
 import org.apache.commons.io.FileUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -19,52 +20,62 @@ public class UploadService {
     private static final Logger LOG = LoggerFactory.getLogger(UploadService.class);
     public static final int DEFAULT_FILE_NAME_LENGTH = 100; //文件名最长长度
 
+    // 格式化上传路径
     private String formatUploadPath(String uploadDir) {
-        uploadDir = uploadDir.replace("\\", "/");
-        if (File.separator.equals(uploadDir.substring(uploadDir.length() - 1))) {
-            return uploadDir;
+        uploadDir = uploadDir.trim().replace("\\", "/");
+        if (uploadDir.length() != 0) {
+            if (File.separator.equals(uploadDir.substring(uploadDir.length() - 1))) {
+                return uploadDir;
+            }
         }
         return uploadDir + File.separator;
     }
 
-    public int checkBigFile(String uploadDir, String fileName, String fileMd5) {
-        uploadDir = formatUploadPath(uploadDir);
+    // 检查文件是否存在chunks
+    public int checkFile(UploadFileInfo uploadFileInfo) {
+        String uploadDir = formatUploadPath(uploadFileInfo.getUploadDir());
 //        // 不覆盖原文件
 //        File file = new File(uploadDir + fileName);
 //        if (file.exists()) {
 //            return -1; //文件已存在，下标为-1
 //        }
         // 读取目录里的所有文件
-        File fileMd5Dir = new File(uploadDir + fileMd5);
+        File fileMd5Dir = new File(uploadDir + uploadFileInfo.getFileMd5());
         File[] chunks = fileMd5Dir.listFiles();
         // 文件未上传过为0，上传中断过返回当前下标
         return chunks == null ? 0 : chunks.length - 1;
     }
 
-    public String upload(String uploadDir, MultipartFile multipartFile, String fileMd5, String chunkIndex) throws Exception {
-        uploadDir = formatUploadPath(uploadDir);
+    // 上传
+    public String upload(MultipartFile multipartFile, UploadFileInfo uploadFileInfo) throws Exception {
+        String uploadDir = formatUploadPath(uploadFileInfo.getUploadDir());
         int fileNamelength = multipartFile.getOriginalFilename().length();
         if (fileNamelength > DEFAULT_FILE_NAME_LENGTH) {
             throw new Exception("[" + fileNamelength + "]文件名超过最大长度：" + DEFAULT_FILE_NAME_LENGTH);
         }
         File fileChunk;
+        String chunkIndex = uploadFileInfo.getChunk();
+        String fileMd5 = uploadFileInfo.getFileMd5();
         if (StringUtils.hasText(chunkIndex)) {
+            // 分片上传
             fileChunk = getAbsoluteFile(uploadDir + fileMd5, chunkIndex);
             FileUtils.copyInputStreamToFile(multipartFile.getInputStream(), fileChunk);
         } else {
+            // 不分片上传
             fileChunk = getAbsoluteFile(uploadDir, multipartFile.getOriginalFilename());
             multipartFile.transferTo(fileChunk);
         }
         return fileChunk.getAbsolutePath();
     }
 
-    public boolean fileMeger(String uploadDir, String fileName, String fileMd5) throws Exception {
-        uploadDir = formatUploadPath(uploadDir);
+    // 分片合并
+    public boolean fileMeger(UploadFileInfo uploadFileInfo) throws Exception {
+        String uploadDir = formatUploadPath(uploadFileInfo.getUploadDir());
         FileChannel outChannel = null;
         File chunksDir = null;
         try {
             // 读取目录里的所有文件
-            chunksDir = new File(uploadDir + fileMd5);
+            chunksDir = new File(uploadDir + uploadFileInfo.getFileMd5());
             File[] chunks = chunksDir.listFiles();
             if (Objects.isNull(chunks) || chunks.length == 0) {
                 return false;
@@ -81,7 +92,7 @@ public class UploadService {
                 }
             });
             // 合并文件
-            File outputFile = getAbsoluteFile(uploadDir, fileName);
+            File outputFile = getAbsoluteFile(uploadDir, uploadFileInfo.getFileName());
             outChannel = new FileOutputStream(outputFile).getChannel();
             FileChannel inChannel = null;
             try {
@@ -121,14 +132,17 @@ public class UploadService {
         return true;
     }
 
+    // 删除分片临时文件及文件夹
     public boolean deleteChunksDir(String uploadDir, String fileMd5) throws Exception {
         uploadDir = formatUploadPath(uploadDir);
         try {
             File chunksDir = new File(uploadDir + fileMd5);
             File[] chunks = chunksDir.listFiles();
-            for (File chunk : chunks) {
-                // 删除所有chunk
-                chunk.delete();
+            if (chunks != null) {
+                for (File chunk : chunks) {
+                    // 删除所有chunk
+                    chunk.delete();
+                }
             }
             // 删除文件夹
             if (chunksDir != null && chunksDir.exists()) {
